Hooked up command parsing to gui, refactored gui initialization to main()
Due to Rust's ownership system, the variables created in app.load() did not live long enough to the end of main()
and the compiler had issues resolving their lifetimes, especially  when running app methods inside an app method.

Thus, as much as it pains me to do so, all the gui initialization code has been moved to main().
This solved the ownership problems with executing commands with a struct method, and the relevant commands
should be able to be successfully executed from the app struct.
Furthermore, the app can now call on itself to change the text in the messagebox.

I will consider moving the adding callback code to a method in the app in the future.