Accounted for panics in certain input cases
When running a command that required input enclosed in double quotes, the app would panic if it received improperly enclosed input (i.e. One double quote was used).
Guard clauses were added to prevent the app from panicking when no arguments were supplied, but it still did not account for the above case.
The regex closure that ran on the command text to capture album name input returned the argument vector if it did not find any, but that was unncessary,
as all commands that needed to capture double quoted text would not work without it.

Thus, the closure was modified to return an Option enum wrapping the result, and the commands would automatically return an error if the closure returned None.
The guard clauses were still kept to provide an appropriate error message (NotEnoughArgs) if ran without any arguments.