Added refreshing for collection screen
Refreshing for the collection screen has been added in Collection.refresh(). This is called after running update and live-refreshes the screen to reflect the new collection.

However, implementing the same thing for wantlist is more difficult as it is passed to a closure on app initialization as a clone, and cannot be refreshed.
Possible solutions include wrapping the wantlist struct in a Rc or RefCell for multiple (mutable) ownership, but the main focus now is to implement the database API and Discogs querying.