Implemented live updating of app data
Changed the method used for the command window in main line 36 to accept a FnMut closure.
This allows app to be passed to it as a mutable variable, and thus can change its own internal data.
Thus, the app does not need to write to disk and reload its data on restart.

I was agonizing over this problem for 5 days, and I solved it in three seconds by simply reading the documentation.
I am actually fucking stupid.
Don't be like me kids. Read the documentation. Read it.