Refactored code for SelectView to store release data
Previously, the SelectView that showed releases stored its data as a vector of Strings.
This would make accessing the album data to display as a popup more computationally expensive, and possibly impossible due to Rust's ownership rules, which are a pain to figure out.

The method to add data to the SelectView was changed to with_all() and add_all(), so as to add the entire Vec of releases into the SelectView itself.
This allowed the data to be passed directly to a builder function to display the album information.