Set up framework for querying Discogs API
The current state of the request file is the framework for querying the Discogs API.
Url and Client builder functions abstract the logic from the query function.
Right now, the code is untested and can only query the main collection folder, but the logic to query all the folders is in work.