Hooked up HTTP request code to main app
The code to make the HTTP request to the Discogs API is complete and has been integrated with the main command code in impl_app.rs.
This also integrates error handling and the displaying of relevant error messages to the messagebox (e.g. 404 error, parsing error, etc)

However, the business logic has not been completed, and the actual parsing of the user's entire collection has yet to be implemented.
This encompasses the procedure of enumerating the entire collection of folders, and parsing each folder, while also accounting for pagination.

Future upgrades include turning the query function into a non-blocking async function.