Hooked up update to read from response
Previously, the update command simply re-read from the pre-saved json file, as the database API was (and still is) not yet complete.
However, now the update command actually modifies the app's internal collection saved in memory with the contents from the web query response.
This is just a proof of concept, and will change again in the future when the database API is complete.

Eventually, the update function will commit changes directly to the database, and the refresh function will read from the database.