Refactored wantlist out of App struct
Ownership and borrowing issues meant that App fields could not be borrowed or partially moved, i.e. to build the wantlist screen.

The code was refactored such that the main app initializes only with the collection data, and the wantlist data is only read
when the callback is triggered and the code to display the screen is run.

This isn't the best solution, and doesn't scale well for users with large wantlists, but then again Discogs' API request throttling also causes this issue.