Added popup screen on album select
This adds a layer to the screen displaying the album information.
As of now this commit contains a working popup, but far from the final appearance.
This was just to ensure that Rust's ownership rules allowed passing the release struct into the build function.

In the future, the build function will return a properly designed screen to display album information.