Implemented iterative querying of Discogs API
The fullupdate() function finally implements a loop to query large collections.
The entire function is not fully complete yet, as it still needs to query the user profile and iterate over the folders to query everything,
but this should be implemented in the near future.

In addition, the Profile struct was created to store the user profile, and to be used in the profile screen (src/screens/profile.rs).