Finalized appearance of albuminfo popup
Updated the appearance of the albuminfo popup to its final appearance.
Only thing left is to properly format the Labels and Formats fields, but that should be trivial.
albuminfo() will also be used by :query to display its results.

The popup.rs file will also be used for the screen displayed when :query returns multiple results.
Upon selection, albuminfo() will be called to display the album information.