Implemented Folders struct
The folders struct is basically a wrapper around a HashMap, but with a specific type specifier and only exposing two main functions:
get() and insert(). This also allows for easier code readability; instead of seeing a HashMap with an extremely long type specifier, the Folder struct can be used instead.