Implemented folder querying and deserialization
The fullupdate() function now queries and the Discogs API for folder metadata and deserializes it into a HashMap
Key: folder name, Value: folder url
The actual query function can then iterate over this hashmap and query the API for each.

The :update command (which calls this function) was tested and returned Ok, which should means that the deserialization of the folder metadata json was successful.
The function would panic if any of the deserialization steps failed (due to all the unwraps).
However, the function is still largely untested and further commits in the future should fix any bugs that may exist.
In addition, fullupdate() should eventually return a Folders struct, and then nothing at all, because everything is written to the database.

Following this, fullupdate() should also handle parsing of the user's wantlist.