Implemented displaying titles for wantlist
The wantlist now can display the proper albums and also pull up the popup that shows album info.
Right now, as always, it still reads from a local file, but that should change once the database API is implemented.

The request::parse_releases() function was modified to accept a ParseType enum to determine what sort of text it needs to parse.
Those changes were committed separately.