Modified parse_releases() to accept ParseType
parse_releases() can now parse both wantlist and collection json, by accepting a ParseType enum and modifying its approach based on that enum.
Some variants in ParseType were unapplicable to this (Initial and Folder), so by default if passed those, it will parse collection json.

However, this function is not user_facing, and it should never be passed those variants in common use.