App now reads token from file
Previously, the Discogs token was read from a file by request::build_client().
This was not reflective of the final design, where the token would be read into the App struct at startup.
This was quickly rectified; the App struct now reads the discogs_token file upon initialization and passes its token when :update is called.