Extracted fullupdate() into separate file
The request file was getting too large (almost 300 loc), and fullupdate() was set to grow even more as the database function calls are added in.
Extracting it into its own separate file gives it space for this growth.

In addition, fullupdate() interfaces with both the database and the query APIs, so it didn't really fit in the request module to begin with.

The code was also cleaned up to change some unwrap() functions to unwrap_or(), setting the variables to Value::Null if unwrap() returned None.
A guard clause following the assignment would immediately return an error if it detected this value.