Added wantlist parsing to fullupdate()
fullupdate() now pulls and parses the user's wantlist as well.
Discogs' pagination also applies to the wantlist, so the same loop had to be applied to pulling the wantlist data.

Thus, the loop was extracted into a private fn get_full() which handles the looping process and returns the data in the form of a Vec<Release>.
It accepts a ParseType so it can handle both the collection and the wantlist, as it calls request::parse_releases(ParseType) internally.