Added database initialization fn
database.rs now contains the new admin module, for general functions like checking the database integrity or initializing the database.
These functions will be called on app startup to check that the required tables are there, and the app will throw up an error message and refuse to start if the checks fail.

As of now the module only contains the init_db function to create the database on first time startup or after a purge, but functions to check integrity and purge the db will be added.