Extracted fullupdate to separate functions
Each aspect of the functionality of fullupdate() has been extracted into its own separate function. This is to support updating each part of the database separately.
The :update command will also be updated to allow switches -c (collection), -p (profile) and -w (wantlist) to expose this functionality to the user.

Additionally, the functions were changed to use the ? notation to return errors and better for error propagation back through the function calls.