Implemented update module of database API
The update module of the database API is (almost) complete. The provided functions can write to the database without any errors, and those that do occur are caught.
The remaining functions to implement are the ones that insert logs and update the user id and token, but those should be easy to implement.

The next module to implement is the query module, and the rest of the admin module (check integrity and purge folders).