Implemented core database functionality
The database API now can update and query, returning the correct data structures.
Most have been implemented except for profile, which should be easy to implement.

Issue: folder tables are dependencies of the folders table, and if the folders table data is altered in any way, the entire database will become unusable.
If data is added, the integrity check will fail.
If data is deleted, the folders whose entries were deleted will be orphaned.
If existing data is altered, the integrity check will fail AND folders will be orphaned.

The current proposed fix is to call the integrity check at the start of fullupdate(), and purge::complete() if it fails.
This should work, but needs testing to ensure it works, and also does not account for orphaned folders.
The integrity check should be changed to fix that.