Added integrity check to fullupdate
fullupdate now performs an integrity check on the database before starting the update. If it finds any errors, it immediately purges the database before updating.
This fixes most of the errors, but does not account for orphan folders, which is currently an outstanding issue.

Furthermore, support for calling from the commandline was added.