Implemented folders in collection screen
The left view in the collection screen now shows the folders present in the database. Selecting an item will change the right view to reflect the folder's contents.
This has been a long time coming and I'm so fucking happy to have finally implemented it.

Possible TODO: Refactor the folders struct to wrap around a BTreeMap instead of a HashMap so as to preserve order, but I'm terribly lazy and I have other parts of the app to work on.