Implemented pre-checks for database
App::initialize() now runs checks to: a. check the database is present and b. check the database integrity.
It will run update::full() if either of those conditions are not met.

The last check to be implemented is the db-config username match check, but that requires config file parsing to be implemented first.