Added datetime support for Release and Profile
Added chrono as a dependency. This will allow for the app to prettily display dates and times.
One issue at the moment is storing the datetime as a String within the database, and not making use of rusqlite's inbuilt ToSql support for chrono's datetime structs.

However, making use of this requires a lot of conversion between timezones, as ToSql and FromSql is only implemented for DateTime<Utc>, while I use DateTime<FixedOffset>.
Will see if this needs to be fixed in the future, but right now datetimes are converted to Strings, since Sqlite3 does that internally anyway.