Updated database to use UTC instead of FixedOffset
Timezone translation now happens at the deserialization level, and all data structures and the database now store time in UTC.
This allows the time to be easily translated to the user's timezone, and enables the app to easily adjust to timezone changes, requiring the user to update their timezone in the config file.

However, writing to the database still requires conversion to strings. To write the data structures natively, it must be split up into multiple SQL statements.
Will see if this is worth doing in the future.