Added loading user-defined timezone
The user timezone can now be set in the config file.
Config now has an associated function timezone() that loads the timezone from the config file and returns it as a FixedOffset.
Popup and profile call this function and use the FixedOffset to offset the time field in the data that they process (which is stored as UTC).