Implemented sparkview drawing for listening history
The history screen now draws sparkviews for each album in the collection, and can be displayed by pressing 4 while in the app.
The keybind to display history has been moved to 'h'.

Issues:
The sparkview loads from the listenlog table and involves a lot of computation to reform the data into something representable by the sparkview,
including padding with unused dates and looping over multiple data structures.
This possibly might not scale well for larger collections and/or larger histories.

Possible solutions include directly storing the data as its own separate table in the database, but that would entail dynamically generating columns for each album,
potentially resulting in a table with up to 1000 columns, and also taking up excess space that is wasted on storing 0s.
Furthermore, updating the collection could possibly mean regenerating the entire table, which means lost data for the end user.

Thus, the only way forward seems to be optimising the conversion algorithm to take up less memory and CPU cycles.
However, this will work for now, and for most collections that aren't too large.