Edited update::full() to write only after pulling
update::full() now only writes to the database after all the data has been pulled from Discogs.
Previously, update::full() would pull and write, and pull and write for each part of the data.
However, now by collating all the data first before writing, it is able to gracefully exit and return an error
if querying Discogs fails for some reason, without affecting the database.

Querying Discogs is the most unreliable part of the update process, and should be completely finished before attempting to write to the database.