Fixed crash on sparkview build if listenlog was empty
A potential bug would cause the app to crash if the listengraph BTreeMap was empty, because it called unwrap() which panics on None.
The unwrap() was replaced with unwrap_or() and would assign Now.date() to first_date.
This would allow the loops to proceed as normal, but only running through once, and displaying only an empty screen, without crashing.

Also added a rev() to the Vec processing chain for the listen history, so as to display the most recent entry at the top.