Added size() to return size of folder
When pushing to a Vec, if its len exceeds its capacity, it will be reallocated in memory, which is expensive and wastes time.
The function calls for creating new Vecs of releases were altered to call with_capacity(), so as to reserve enough space for the entire set without reallocating.
This should save some time, especially when pulling larger collections from the database.