Converted all text to be ascii-compatible
A new field search_string was added to the Release struct as a dedicated search match term, instead of matching to the album title.
This converted the title to a fully ascii-compatible string, allowing searches to use entirely ascii characters to query the database.
A clause was added to remove all non-alphanumeric characters from the search string, and all the queries that contained them, which meant that searches could optionally use them and still succeed either way.