Added multithreading to collection update
The update::full() function now uses multithreading to call all folders concurrently. This won't see a large performance improvement with smaller collections,
and might even worsen slightly due to cloning the Client struct for each folder, but for larger collections there should be an improvement.
The Discogs API is paginated, and there is no way to multithread for that.

Might experiment with mpsc instead of joining in the future.