Moved updating wantlist into a separate thread
The wantlist update process now runs in a separate thread. This is to guard against json parsing errors, where the unwraps
in the json parsing function will cause the thread to panic.
The join handle will return an Err if the thread panics, and will bubble up through the call stack, instead of crashing out,
if the update wantlist function is run in the main thread.
However, the update thread is synchronous with the main update function, so no time is saved.

In the future, I plan to implement the wantlist update function to run concurrently alongside the collection update function,
which should save some more time.