Bugfix on json parsing
One of the fields on the json may or not be present, depending on the album.
The ok_or() change was applied to deserializing this field, and thus caused the app to exit with an error when updating.
This was quickly reverted back to unwrap_or().

Also in this commit, some superfluous code in the main update loop was removed, now that From was implemented and ? could be used
instead of explicitly matching on errors.
Also, some unwraps that were missed were quickly replaced with ok_or().