Fixed db bug with duplicate entries
Using Discogs ID as a primary key for the database did not work when
multiple entries of the same release were in the database.

Thus, a new primary key was used; an index. Each entry into the database
is assigned a sequential integer as it is added to the database during
an update, and is guaranteed to be unique within the table it is added
to.

Just to be safe, hashing support was added and can be swapped to if
needed, and a separate column was added for a hash digest of each entry.